Software/requirements.txt
matplotlib==3.3.4
numpy==1.20.1
PyQt5==5.15.2
python-box==5.3.0
PyYAML==5.4.1
PyQt5==5.15.2
pyqt5-tools==5.15.2.3.0.2
pyrtlsdr==0.2.92
scipy
samplerate
pipwin

#pipwin install simpleaudio


